refactor(urls): move url constants to app/renderer
